+++ src/payload/nw4r/snd/DvdFileStream.cpp
@@ -5,7 +5,7 @@ namespace snd {
 
 DvdSoundArchive::DvdFileStream::DvdFileStream(s32 entryNum, u32 offset, s32 size):
     ut::DvdLockedFileStream(entryNum),
-    offset(offset),
+    offset((int)offset),
     size(size) {
 
     ut::DvdLockedFileStream::Seek(this->offset, ut::FILE_STREAM_SEEK_BEGIN);
@@ -13,7 +13,7 @@ DvdSoundArchive::DvdFileStream::DvdFileStream(s32 entryNum, u32 offset, s32 size
 
 DvdSoundArchive::DvdFileStream::DvdFileStream(const DVDFileInfo* fileInfo, u32 offset, s32 size):
     ut::DvdLockedFileStream(fileInfo, false),
-    offset(offset),
+    offset((int)offset),
     size(size) {
 
     ut::DvdLockedFileStream::Seek(this->offset, ut::FILE_STREAM_SEEK_BEGIN);
